pyproject.toml
[project]
name = "townspark-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-nested-routers>=0.94.1",
    "pillow>=11.0.0",
]

[dependency-groups]
dev = [
    "requests>=2.32.5",
]
